Refactor the renderWithRedux util
