+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui_v2/src/__tests__/MainLayout.spec.tsx
@@ -1,11 +1,8 @@
-import { configureStore } from '@reduxjs/toolkit';
-import { render as rtlRender } from '@testing-library/react';
 import React from 'react';
-import { Provider } from 'react-redux';
 
 import { MainLayout } from '../components/MainLayout';
-import { rootReducer } from '../store/rootReducer';
 import { initialState as initialSetting } from '../store/setting/settingReducer';
+import { render } from '../testUtils/renderWithRedux';
 
 jest.mock('react-router-dom', () => ({
   ...jest.requireActual('react-router-dom'),
@@ -16,41 +13,26 @@ jest.mock('react-router-dom', () => ({
 
 jest.mock('axios');
 
-function render(
-  ui,
-  {
-    initialState = undefined,
-    store = configureStore({ reducer: rootReducer, preloadedState: initialState }),
-    ...renderOptions
-  } = {},
-) {
-  function Wrapper({ children }) {
-    return <Provider store={store}>{children}</Provider>;
-  }
-  return {
-    ...rtlRender(ui, { wrapper: Wrapper, ...renderOptions }),
-    store,
-  };
-}
+describe('Setting Panel Behaviour', () => {
+  it('The setting panel should be open if `isTranerValid` is false.', () => {
+    const { getByText } = render(<MainLayout />, {
+      initialState: { setting: { ...initialSetting, isTrainerValid: false } },
+    });
 
-test('The setting panel should be open if `isTranerValid` is false.', () => {
-  const { getByText } = render(<MainLayout />, {
-    initialState: { setting: { ...initialSetting, isTrainerValid: false } },
+    expect(getByText(/Settings/)).not.toBeNull();
   });
 
-  expect(getByText(/Settings/)).not.toBeNull();
-});
-
-test("The project dropdown shouldn't be shown when `isTrainerValid` is false", () => {
-  const { queryByText } = render(<MainLayout />, {
-    initialState: { setting: { ...initialSetting, isTrainerValid: false } },
+  it("The project dropdown shouldn't be shown when `isTrainerValid` is false", () => {
+    const { queryByText } = render(<MainLayout />, {
+      initialState: { setting: { ...initialSetting, isTrainerValid: false } },
+    });
+    expect(queryByText(/Project/)).toBeNull();
   });
-  expect(queryByText(/Project/)).toBeNull();
-});
 
-test('The data policy dialog should be shown when the app is initialized', () => {
-  const { getByText } = render(<MainLayout />, {
-    initialState: { setting: { ...initialSetting, isTrainerValid: false, appInsightHasInit: false } },
+  it('The data policy dialog should be shown when the app is initialized', () => {
+    const { getByText } = render(<MainLayout />, {
+      initialState: { setting: { ...initialSetting, isTrainerValid: false, appInsightHasInit: false } },
+    });
+    expect(getByText(/Data Collection Policy/)).not.toBeNull();
   });
-  expect(getByText(/Data Collection Policy/)).not.toBeNull();
 });
