+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui_v2/src/testUtils/renderWithRedux.tsx
@@ -1,0 +1,22 @@
+import { render as rtlRender } from '@testing-library/react';
+import { configureStore } from '@reduxjs/toolkit';
+import React from 'react';
+import { Provider } from 'react-redux';
+import { rootReducer } from '../store/rootReducer';
+
+export function render(
+  ui,
+  {
+    initialState = undefined,
+    store = configureStore({ reducer: rootReducer, preloadedState: initialState }),
+    ...renderOptions
+  } = {},
+) {
+  function Wrapper({ children }) {
+    return <Provider store={store}>{children}</Provider>;
+  }
+  return {
+    ...rtlRender(ui, { wrapper: Wrapper, ...renderOptions }),
+    store,
+  };
+}
